Fix typo in JsonValue which broke the conversion to uint and add test case
(cherry picked from commit 314568184ca6bf4dbe5e8dcca362747225de3102)

+++ mcs/class/System.Json/System.Json/JsonValue.cs
@@ -442,7 +442,7 @@ namespace System.Json
 		{
 			if (value == null)
 				throw new ArgumentNullException ("value");
-			return Convert.ToUInt16 (((JsonPrimitive) value).Value, NumberFormatInfo.InvariantInfo);
+			return Convert.ToUInt32 (((JsonPrimitive) value).Value, NumberFormatInfo.InvariantInfo);
 		}
 
 		public static implicit operator ulong (JsonValue value)
